+++ gem/test/test_helper.rb
@@ -90,6 +90,8 @@ module SolidGcp
       SolidGcp.config.project = "test-project"
       SolidGcp.config.cable.mode = :off
       SolidGcp.config.cable.project = nil
+      SolidGcp.config.cable.firestore_emulator_host = nil
+      SolidGcp.config.cable.auth_emulator_host = nil
       SolidGcp.config.cable.touch_debounce = 1.second
       SolidGcp::Cable.reset_missing_project_warning!
       SolidGcp::Cable::TestSink.clear!